refactor: co-locate effect registrations with resolvers
Move effect handler registrations from centralized effectRegistrations.ts
into their respective module files. Each module now exports a
registerXxxEffects() function that registers its handlers.

- Create effectHelpers.ts with shared getPlayerContext() helper
- Add registration functions to 13 effect modules
- Simplify effectRegistrations.ts from 443 to 90 lines as coordinator
- Export registration functions from index.ts

Benefits:
- Registration logic lives with resolution logic
- Easier maintenance - adding effects requires one file
- Better discoverability - each module is self-contained
